# docs/g0/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Hoa Nam Public Product Viewer API
  version: 1.0.0-g0
  description: |
    G0 public-contract baseline only. All product reads are limited at the server boundary
    to PUBLISHED records with approved public availability IN_STOCK or PREORDER. PREORDER is a
    requestable public state only; it contains no inventory quantity or internal stock status.
    This API contains no prices, inventory quantities,
    internal catalogue data, CRM/ERP integration, outbox, queue, retry worker, notification,
    or Lead-management lifecycle.
servers:
  - url: https://api.example.invalid
    description: Placeholder only; environment values are configured outside this document.
...
paths:
  /api/v1/public/config:
    get:
      tags: [public-config]
      operationId: getPublicConfig
      summary: Get public operational configuration without secrets.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Public config retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigEnvelope'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/public/home:
    get:
      tags: [catalogue]
      operationId: getHome
      summary: Get public home domains and non-empty content sections.
      description: Empty or unavailable featured/category sections are omitted; clients must not render empty sections.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Home payload retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeEnvelope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/public/quote-requests:
    post:
      tags: [quote]
      operationId: createQuoteRequest
      summary: Validate and minimally persist one public quote request.
      description: |
        This endpoint accepts a currently public IN_STOCK or PREORDER product/variant and stops after
        minimal persistence and acknowledgement. It does not create an order,
        CRM/ERP record, notification, queue message, outbox event, retry job, or lifecycle state.
        The source is server-derived as ZALO_MINI_APP; it is not accepted in the body.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequestInput'
      responses:
        '201':
          description: Minimal quote record created and acknowledged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteAcceptedEnvelope'
        '200':
          description: Equivalent replay of an already accepted idempotent request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteAcceptedEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Product availability changed or idempotency key has a different payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  parameters:
    XRequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Optional non-secret client correlation ID. The server always returns its request ID in meta.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque 16–128 character submission key. Reuse only for a retry of the same normalized quote payload.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._~-]+$'
...
  responses:
    BadRequest:
      description: Validation, cursor, or filter error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Request rate exceeded. meta.retry_after_seconds, when supplied, is authoritative.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailable:
      description: Maintenance or safe upstream unavailability state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    EnvelopeBase:
      type: object
      required: [success, message, data, meta, error_code, errors]
      properties:
        success:
          type: boolean
        message:
          type: string
          maxLength: 280
        data: {}
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error_code:
          type: [string, 'null']
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        errors:
          oneOf:
            - $ref: '#/components/schemas/FieldErrors'
            - type: 'null'
    ResponseMeta:
      type: object
      additionalProperties: false
      properties:
        request_id:
          type: string
          minLength: 1
          maxLength: 128
        data_version:
          type: string
          format: date-time
        next_cursor:
          type: [string, 'null']
          maxLength: 1024
        limit:
          type: integer
          minimum: 1
          maximum: 50
        retry_after_seconds:
          type: integer
          minimum: 1
        estimated_end_at:
          type: string
          format: date-time
        idempotent_replay:
          type: boolean
    FieldErrors:
      type: object
      additionalProperties:
        type: array
        minItems: 1
        items:
          type: string
          maxLength: 280
    ErrorCode:
      type: [string, 'null']
      enum:
        - VALIDATION_ERROR
        - INVALID_CURSOR
        - INVALID_FILTER
        - PRODUCT_NOT_FOUND
        - PRODUCT_NOT_AVAILABLE
        - IDEMPOTENCY_CONFLICT
        - RATE_LIMITED
        - MAINTENANCE
        - UPSTREAM_UNAVAILABLE
        - INTERNAL_ERROR
        - null
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            success:
              const: false
            data:
              type: 'null'
            error_code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_CURSOR
                - INVALID_FILTER
                - PRODUCT_NOT_FOUND
                - PRODUCT_NOT_AVAILABLE
                - IDEMPOTENCY_CONFLICT
                - RATE_LIMITED
                - MAINTENANCE
                - UPSTREAM_UNAVAILABLE
                - INTERNAL_ERROR
...
    ContactPhone:
      type: object
      required: [display, tel]
      additionalProperties: false
      properties:
        display:
          type: string
          minLength: 1
          maxLength: 80
        tel:
          type: string
          pattern: '^\+[1-9]\d{7,14}$'
    ZaloOa:
      type: object
      required: [id, chat_url]
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 128
        chat_url:
          type: string
          format: uri
    SupportHours:
      type: object
      required: [timezone, intervals]
      additionalProperties: false
      properties:
        timezone:
          type: string
          const: Asia/Ho_Chi_Minh
        intervals:
          type: array
          items:
            type: object
            required: [days, opens_at, closes_at]
            additionalProperties: false
            properties:
              days:
                type: array
                minItems: 1
                items:
                  type: string
                  enum: [MON, TUE, WED, THU, FRI, SAT, SUN]
              opens_at:
                type: string
                pattern: '^([01]\d|2[0-3]):[0-5]\d$'
              closes_at:
                type: string
                pattern: '^([01]\d|2[0-3]):[0-5]\d$'
    Maintenance:
      type: object
      required: [enabled, message, estimated_end_at]
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        message:
          type: [string, 'null']
          maxLength: 280
        estimated_end_at:
          type: [string, 'null']
          format: date-time
    PublicConfig:
      type: object
      required:
        - config_version
        - hotline
        - zalo_oa
        - support_hours
        - privacy_policy_url
        - maintenance
        - min_supported_app_version
        - feature_flags
      additionalProperties: false
      properties:
        config_version:
          type: string
          minLength: 1
          maxLength: 128
        hotline:
          oneOf:
            - $ref: '#/components/schemas/ContactPhone'
            - type: 'null'
        hotline_fallback:
          oneOf:
            - $ref: '#/components/schemas/ContactPhone'
            - type: 'null'
        zalo_oa:
          oneOf:
            - $ref: '#/components/schemas/ZaloOa'
            - type: 'null'
        support_hours:
          oneOf:
            - $ref: '#/components/schemas/SupportHours'
            - type: 'null'
        privacy_policy_url:
          type: [string, 'null']
          format: uri
        maintenance:
          $ref: '#/components/schemas/Maintenance'
        min_supported_app_version:
          type: [string, 'null']
          maxLength: 80
        feature_flags:
          type: object
          additionalProperties:
            type: boolean
...
    HomePayload:
      type: object
      required: [domains, sections]
      additionalProperties: false
      properties:
        domains:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: '#/components/schemas/Domain'
        sections:
          type: array
          items:
            $ref: '#/components/schemas/HomeSection'
    QuoteRequestInput:
      type: object
      required: [product_id, full_name, phone, consent, privacy_version]
      additionalProperties: false
      properties:
        product_id:
          type: string
          minLength: 1
          maxLength: 128
        variant_id:
          type: [string, 'null']
          maxLength: 128
        full_name:
          type: string
          minLength: 2
          maxLength: 80
        phone:
          type: string
          pattern: '^(?:0(?:3|5|7|8|9)\d{8}|\+84(?:3|5|7|8|9)\d{8})$'
        province_code:
          type: [string, 'null']
          maxLength: 32
          pattern: '^[A-Z0-9_-]+$'
        note:
          type: [string, 'null']
          maxLength: 500
        consent:
          type: boolean
          const: true
        privacy_version:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[ -~]+$'
    QuoteAccepted:
      type: object
      required: [request_id, status]
      additionalProperties: false
      properties:
        request_id:
          type: string
          minLength: 1
          maxLength: 128
        status:
          type: string
          const: RECEIVED
...
    ConfigEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            success: { const: true }
            data: { $ref: '#/components/schemas/PublicConfig' }
            error_code: { type: 'null' }
            errors: { type: 'null' }
    HomeEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            success: { const: true }
            data: { $ref: '#/components/schemas/HomePayload' }
            error_code: { type: 'null' }
            errors: { type: 'null' }
...
    QuoteAcceptedEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            success: { const: true }
            data: { $ref: '#/components/schemas/QuoteAccepted' }
            error_code: { type: 'null' }
            errors: { type: 'null' }
...